pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-template-pro"
version = "0.1.0"
description = "Template FastAPI template-ready"
readme = "README.md"
requires-python = ">=3.11,<3.13"
keywords = ["fastapi", "template", "api"]
authors = [{ name = "Sereina Yopa" }]

dependencies = [
  "fastapi==0.116.2",
  "starlette==0.48.0",
  "uvicorn[standard]==0.35.0",
  "pydantic==2.11.9",
  "pydantic-settings==2.10.1",
  "sqlalchemy==2.0.43",
  "asyncpg==0.30.0",
  "alembic==1.16.5",
  "httpx==0.28.1",
  "email-validator==2.3.0",
  "python-jose[cryptography]==3.5.0",
  "passlib[bcrypt]==1.7.4",
  "python-dotenv==1.1.1",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.2",
  "coverage[toml]==7.10.6",
  "ruff==0.13.0",
  "black==25.1.0",
  "flake8==7.3.0",
  "mypy==1.18.1",
  "pre-commit==4.3.0",
  "types-passlib==1.7.7.20250602",
]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
strict = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["app/tests"]
